+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Viash 0.x.x (yyyy-MM-dd): TODO Add title
 
+## NEW FUNCTIONALITY
+
+* `viash test` and `viash ns test`: add `--setup` argument to determine the docker build strategy before a component is tested.
+
 ## MINOR CHANGES
 
 * `PythonScript`: Pass `-B` to Python to avoid creating *.pyc and *.pyo files on importing helper functions (PR #442).